refactor: split linux metadata guard runtime
Move protected metadata guard target modeling, runtime supervision, and cleanup/watch logic into focused Linux sandbox modules.
