+++ codex-rs/linux-sandbox/src/lib.rs
@@ -6,11 +6,19 @@
 #[cfg(target_os = "linux")]
 mod bwrap;
 #[cfg(target_os = "linux")]
+mod bwrap_runtime;
+#[cfg(target_os = "linux")]
 mod landlock;
 #[cfg(target_os = "linux")]
 mod launcher;
 #[cfg(target_os = "linux")]
 mod linux_run_main;
+#[cfg(target_os = "linux")]
+mod metadata_guard;
+#[cfg(target_os = "linux")]
+mod metadata_guard_watcher;
+#[cfg(target_os = "linux")]
+mod metadata_paths;
 #[cfg(target_os = "linux")]
 mod proxy_routing;
 #[cfg(target_os = "linux")]
